+++ CHANGES.rst
@@ -171,6 +171,10 @@ Changelog
   [thomlamb]
 
 - Fixed faceted map
+- Change some icons : SectionHTML and SectionExternalContent
+  [boulch]
+
+- MWEBTUBA : Add new section : imio.smartweb.SectionExternalContent (Manage embeded contents)
   [boulch]
 
 - WEB-3837 : Can define specific news to get (instead of all news from news folders)